+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 0.5.0 (unreleased)
 
+## Changed
+
+- Exchanged rust-overlay for fenix. When you override your inputs, you now have
+  to override fenix instead of rust-overlay.
+
 ### Removed
 
 - Removed the internal option `boot.lanzaboote.enrollKeys` that was only